Move `Operator` back to `operator.py` (for now)
