+++ nvtabular/ops/__init__.py
@@ -14,8 +14,6 @@
 # limitations under the License.
 #
 
-from .base import Operator
-
 # alias submodules here to avoid breaking everything with moving to submodules
 # flake8: noqa
 from .bucketize import Bucketize
@@ -35,6 +33,7 @@ from .lambdaop import LambdaOp
 from .list_slice import ListSlice
 from .logop import LogOp
 from .normalize import Normalize, NormalizeMinMax
+from .operator import Operator
 from .rename import Rename
 from .stat_operator import StatOperator
 from .target_encoding import TargetEncoding
