+++ nvtabular/ops/clip.py
@@ -16,7 +16,7 @@
 from nvtabular.dispatch import DataFrameType, annotate
 
 from ..column import ColumnSchemas
-from .base import Operator
+from .operator import Operator
 
 
 class Clip(Operator):
